+++ controlplane/src/core/repositories/SchemaCheckRepository.ts
@@ -28,6 +28,7 @@ import {
 } from '../../db/schema.js';
 import {
   CheckedSubgraphDTO,
+  COMPOSITION_IGNORE_EXTERNAL_KEYS_FEATURE_ID,
   FederatedGraphDTO,
   Label,
   NamespaceDTO,
@@ -757,6 +758,13 @@ export class SchemaCheckRepository {
       organizationId,
       featureId: 'breaking-change-retention',
     });
+    const ignoreExternalKeys =
+      (
+        await orgRepo.getFeature({
+          organizationId,
+          featureId: COMPOSITION_IGNORE_EXTERNAL_KEYS_FEATURE_ID,
+        })
+      )?.enabled ?? false;
 
     const limit = changeRetention?.limit ?? defaultRetentionLimitInDays;
 
@@ -825,7 +833,12 @@ export class SchemaCheckRepository {
       let newGraphQLSchema: GraphQLSchema | undefined;
       if (newSchemaSDL) {
         try {
-          // Here we check if the schema is valid as a subgraph SDL
+          /* Here we check if the schema is valid as a subgraph SDL
+           * `buildSchema` only calls normalization in isolation.
+           * The `disableResolvabilityChecks` flag is only used in the federation step.
+           * The `ignoreExternalKeys` flag is propagated in normalization but only used in the federation step.
+           * Consequently, there is currently no reason to propagate the options within `buildSchema`.
+           */
           const result = buildSchema(newSchemaSDL, true, routerCompatibilityVersion);
           if (!result.success) {
             await this.update({
@@ -857,7 +870,7 @@ export class SchemaCheckRepository {
           }
           if (namespace.enableGraphPruning) {
             const parsedSchema = parse(newSchemaSDL);
-            // this new GraphQL schema conatins the location info
+            // this new GraphQL schema contains the location info
             newGraphQLSchema = buildASTSchema(parsedSchema, { assumeValid: true, assumeValidSDL: true });
           }
         } catch (e: any) {
@@ -1119,6 +1132,10 @@ export class SchemaCheckRepository {
     }
 
     const { composedGraphs } = await composer.composeWithProposedSchemas({
+      compositionOptions: {
+        disableResolvabilityValidation: false,
+        ignoreExternalKeys,
+      },
       inputSubgraphs: checkSubgraphs,
       graphs: federatedGraphs.filter((g) => !g.contract),
     });
@@ -1422,7 +1439,10 @@ export class SchemaCheckRepository {
         limit: targetLimit,
         chClient,
         newGraphQLSchema: targetNewGraphQLSchema,
-        disableResolvabilityValidation: false,
+        compositionOptions: {
+          disableResolvabilityValidation: false,
+          ignoreExternalKeys,
+        },
         webhookService,
       });