+++ router/pkg/config/config.go
@@ -616,10 +616,19 @@ type NatsAuthentication struct {
 	NatsTokenBasedAuthentication `yaml:"token,inline"`
 }
 
+type NatsDurableConsumersConfiguration struct {
+	DeleteOnShutdown bool `yaml:"delete_on_shutdown" envDefault:"false"`
+}
+
+type NatsConsumersConfiguration struct {
+	Durable NatsDurableConsumersConfiguration `yaml:"durable,omitempty"`
+}
+
 type NatsEventSource struct {
-	ID             string              `yaml:"id,omitempty"`
-	URL            string              `yaml:"url,omitempty"`
-	Authentication *NatsAuthentication `yaml:"authentication,omitempty"`
+	ID             string                     `yaml:"id,omitempty"`
+	URL            string                     `yaml:"url,omitempty"`
+	Authentication *NatsAuthentication        `yaml:"authentication,omitempty"`
+	Consumers      NatsConsumersConfiguration `yaml:"consumers,omitempty"`
 }
 
 func (n NatsEventSource) GetID() string {