+++ router/pkg/trace/config.go
@@ -72,7 +72,7 @@ type Config struct {
 	TestMemoryExporter sdktrace.SpanExporter
 	// TestErrorHandler is used for testing purposes. If set, the handler will be used to
 	// handle exporter errors locally instead of relying on the global otel.SetErrorHandler.
-	TestErrorHandler func(error)
+	TestErrorHandler    func(error)
 	ResponseTraceHeader config.ResponseTraceHeader
 	Attributes          []config.CustomAttribute
 	// SanitizeUTF8 configures sanitization of invalid UTF-8 sequences in span attribute values